+++ quickbi-public-20220101/alibabacloud_quickbi_public20220101/models/_add_user_request.py
@@ -11,6 +11,7 @@ class AddUserRequest(DaraModel):
         account_name: str = None,
         admin_user: bool = None,
         auth_admin_user: bool = None,
+        copilot_modules: str = None,
         nick_name: str = None,
         role_ids: str = None,
         user_type: int = None,
@@ -39,6 +40,7 @@ class AddUserRequest(DaraModel):
         # 
         # <notice>This parameter is deprecated and not recommended for use. It is invalid when RoleIds is provided.</notice>
         self.auth_admin_user = auth_admin_user
+        self.copilot_modules = copilot_modules
         # Aliyun account nickname.
         # 
         # - Format check: Maximum length of 50 characters.
@@ -79,6 +81,9 @@ class AddUserRequest(DaraModel):
         if self.auth_admin_user is not None:
             result['AuthAdminUser'] = self.auth_admin_user
 
+        if self.copilot_modules is not None:
+            result['CopilotModules'] = self.copilot_modules
+
         if self.nick_name is not None:
             result['NickName'] = self.nick_name
 
@@ -103,6 +108,9 @@ class AddUserRequest(DaraModel):
 
         if m.get('AuthAdminUser') is not None:
             self.auth_admin_user = m.get('AuthAdminUser')
+
+        if m.get('CopilotModules') is not None:
+            self.copilot_modules = m.get('CopilotModules')
 
         if m.get('NickName') is not None:
             self.nick_name = m.get('NickName')