+++ quickbi-public-20220101/alibabacloud_quickbi_public20220101/models/_query_user_list_response_body.py
@@ -137,6 +137,7 @@ class QueryUserListResponseBodyResultData(DaraModel):
         account_name: str = None,
         admin_user: bool = None,
         auth_admin_user: bool = None,
+        copilot_modules: List[str] = None,
         is_deleted: bool = None,
         joined_date: int = None,
         last_login_time: int = None,
@@ -159,6 +160,7 @@ class QueryUserListResponseBodyResultData(DaraModel):
         # *   true
         # *   false
         self.auth_admin_user = auth_admin_user
+        self.copilot_modules = copilot_modules
         # User status: 
         # - Active - false 
         # - Inactive - true
@@ -200,6 +202,9 @@ class QueryUserListResponseBodyResultData(DaraModel):
         if self.auth_admin_user is not None:
             result['AuthAdminUser'] = self.auth_admin_user
 
+        if self.copilot_modules is not None:
+            result['CopilotModules'] = self.copilot_modules
+
         if self.is_deleted is not None:
             result['IsDeleted'] = self.is_deleted
 
@@ -236,6 +241,9 @@ class QueryUserListResponseBodyResultData(DaraModel):
 
         if m.get('AuthAdminUser') is not None:
             self.auth_admin_user = m.get('AuthAdminUser')
+
+        if m.get('CopilotModules') is not None:
+            self.copilot_modules = m.get('CopilotModules')
 
         if m.get('IsDeleted') is not None:
             self.is_deleted = m.get('IsDeleted')